+++ src/HCSR04.cpp
@@ -9,7 +9,7 @@
 HCSR04Sensor::HCSR04Sensor() {}
 HCSR04Sensor::~HCSR04Sensor() { this->end(); }
 
-void HCSR04Sensor::begin(int triggerPin, int* echoPins, short echoCount, int timeout, eUltraSonicUnlock_t unlock) {
+void HCSR04Sensor::begin(uint8_t triggerPin, uint8_t* echoPins, uint8_t echoCount, int timeout, eUltraSonicUnlock_t unlock) {
 	if (this->echoCount != echoCount) this->end();
 	
 	this->triggerPin = triggerPin;
@@ -24,12 +24,12 @@ void HCSR04Sensor::begin(int triggerPin, int* echoPins, short echoCount, int tim
 	if (this->triggerTimes == NULL) this->triggerTimes = new unsigned long[echoCount];
 	if (this->echoTimes == NULL) this->echoTimes = new unsigned long[echoCount];
 
-	if (this->echoInts == NULL) this->echoInts = new int[echoCount];
-	if (this->echoStages == NULL) this->echoStages = new int[echoCount];
-	if (this->echoMasks == NULL) this->echoMasks = new int[echoCount];
-	if (this->echoPorts == NULL) this->echoPorts = new int[echoCount];
+	if (this->echoInts == NULL) this->echoInts = new uint8_t[echoCount];
+	if (this->echoStages == NULL) this->echoStages = new uint8_t[echoCount];
+	if (this->echoMasks == NULL) this->echoMasks = new uint8_t[echoCount];
+	if (this->echoPorts == NULL) this->echoPorts = new uint8_t[echoCount];
 
-	for (int i = 0; i < this->echoCount; i++) {
+	for (uint8_t i = 0; i < this->echoCount; i++) {
 		this->triggerTimes[i] = 0;
 		this->echoTimes[i] = 0;
 
@@ -92,7 +92,7 @@ void HCSR04Sensor::measureMicroseconds(long* results) {
 	digitalWrite(triggerPin, LOW);
 	
 	// Attach interrupts to echo pins for the starting point
-	for (int i = 0; i < this->echoCount; i++) {
+	for (uint8_t i = 0; i < this->echoCount; i++) {
 		if (this->echoInts[i] >= 0 && this->echoStages[i] == 0) {
 			this->echoStages[i] = 1;
 			switch (i) {
@@ -119,7 +119,7 @@ void HCSR04Sensor::measureMicroseconds(long* results) {
 		currentMicros = micros();
 		elapsedMicros = currentMicros - startMicros;
 
-		for (int i = 0; i < this->echoCount; i++) {
+		for (uint8_t i = 0; i < this->echoCount; i++) {
 			waiting &= elapsedMicros < this->timeout || (this->triggerTimes[i] > 0 && this->echoTimes[i] == 0 && (currentMicros - this->triggerTimes[i]) < this->timeout);
 
 			if (this->triggerTimes[i] == 0 && this->echoPorts[i] >= 0) {
@@ -165,7 +165,7 @@ void HCSR04Sensor::measureMicroseconds(long* results) {
 	}
 	
 	// Determine the durations of each sensor.
-	for (int i = 0; i < this->echoCount; i++) {
+	for (uint8_t i = 0; i < this->echoCount; i++) {
 		if (this->echoInts[i] >= 0) this->echoStages[i] = 0;
 		if (this->triggerTimes[i] > 0 && this->echoTimes[i] > 0) {
 			long resultTime = this->echoTimes[i] - this->triggerTimes[i];
@@ -188,7 +188,7 @@ void HCSR04Sensor::measureDistanceMm(float temperature, double* results) {
 	long* times = measureMicroseconds();
 	
 	// Calculate the distance in mm for each result.
-	for (int i = 0; i < this->echoCount; i++) {
+	for (uint8_t i = 0; i < this->echoCount; i++) {
 		double distanceMm = times[i] / 2.0 * speedOfSoundInMmPerMs;
 		if (distanceMm < 10 || distanceMm > 4000) {
 			results[i] = HCSR04_INVALID_RESULT;
@@ -205,7 +205,7 @@ void HCSR04Sensor::measureDistanceCm(float temperature, double* results) {
 	long* times = measureMicroseconds();
 	
 	// Calculate the distance in cm for each result.
-	for (int i = 0; i < this->echoCount; i++) {
+	for (uint8_t i = 0; i < this->echoCount; i++) {
 		double distanceCm = times[i] / 2.0 * speedOfSoundInCmPerMs;
 		if (distanceCm < 1 || distanceCm > 400) {
 			results[i] = HCSR04_INVALID_RESULT;
@@ -222,7 +222,7 @@ void HCSR04Sensor::measureDistanceIn(float temperature, double* results) {
 	long* times = measureMicroseconds();
 
 	// Calculate the distance in cm for each result.
-	for (int i = 0; i < this->echoCount; i++) {
+	for (uint8_t i = 0; i < this->echoCount; i++) {
 		double distanceIn = times[i] / 2.0 * speedOfSoundInCmPerMs;
 		if (distanceIn < 1 || distanceIn > 157.4804) {
 			results[i] = HCSR04_INVALID_RESULT;
@@ -233,12 +233,12 @@ void HCSR04Sensor::measureDistanceIn(float temperature, double* results) {
 	}
 }
 
-void HCSR04Sensor::unlockSensors(eUltraSonicUnlock_t unlock, int* echoPins) {
+void HCSR04Sensor::unlockSensors(eUltraSonicUnlock_t unlock, uint8_t* echoPins) {
 	if (unlock == eUltraSonicUnlock_t::unlockSkip) return;
 	bool hasLocked = false;
 
 	// Check if any sensor is in a locked state and unlock it if necessary.
-	for (int i = 0; echoPins[i] != 0; i++) {
+	for (uint8_t i = 0; echoPins[i] != 0; i++) {
 		if (unlock == eUltraSonicUnlock_t::unlockMaybe && digitalRead(echoPins[i]) == LOW) continue;
 		pinMode(echoPins[i], OUTPUT);
 		digitalWrite(echoPins[i], LOW);
@@ -248,18 +248,18 @@ void HCSR04Sensor::unlockSensors(eUltraSonicUnlock_t unlock, int* echoPins) {
 	if (hasLocked) delay(100);
 
 	// Revert the pinMode after potential unlocking.
-	for (int i = 0; echoPins[i] != 0; i++) {
+	for (uint8_t i = 0; echoPins[i] != 0; i++) {
 		pinMode(echoPins[i], INPUT);
 	}
 	
 	if (hasLocked) delay(100);
 }
 
-void HCSR04Sensor::triggerInterrupt(int index) {
+void HCSR04Sensor::triggerInterrupt(uint8_t index) {
 	if (this->triggerTimes[index] == 0) this->triggerTimes[index] = micros();
 }
 
-void HCSR04Sensor::echoInterrupt(int index) {
+void HCSR04Sensor::echoInterrupt(uint8_t index) {
 	if (this->triggerTimes[index] > 0 && this->echoTimes[index] == 0) this->echoTimes[index] = micros();
 }
